teacher journey ui and qa issues
