+++ src/Teachers/Aside.js
@@ -182,7 +182,7 @@ const Aside = ({ rtl, toggled, handleToggleSidebar }) => {
                             Manage FAQ&apos;s
                         </NavLink>
                     </MenuItem>
-                    {/* <MenuItem
+                    <MenuItem
                         icon={<FaShieldVirus />}
                         className={
                             (location.pathname === '/teacher/support-journey' || location.pathname === '/teacher/support-journey/add-ticket' )  &&
@@ -197,7 +197,7 @@ const Aside = ({ rtl, toggled, handleToggleSidebar }) => {
                             {' '}
                             Support Journey
                         </NavLink>
-                    </MenuItem> */}
+                    </MenuItem>
                 </Menu>
             </SidebarContent>
         </ProSidebar>
