+++ src/assets/sass/themes/_rdt.scss
@@ -33,3 +33,7 @@
 .bMgaAx div:first-child, .dBbhDz div:first-child{
   white-space: initial !important;
 }
+
+.data-table-extensions>.data-table-extensions-action>button:after{
+  margin-top: 20px !important;
+}
